@@ -36,12 +36,12 @@
     "node-sass": "4.11.0",
     "optimize-css-assets-webpack-plugin": "5.0.1",
     "postcss-loader": "2.1.5",
-    "prop-types": "15.6.2",
+    "prop-types": "15.7.2",
     "raven-js": "3.26.3",
-    "react": "16.4.2",
-    "react-dom": "16.4.2",
-    "react-redux": "5.0.7",
-    "redux": "3.7.2",
+    "react": "16.8.5",
+    "react-dom": "16.8.5",
+    "react-redux": "6.0.1",
+    "redux": "4.0.1",
     "redux-logger": "3.0.6",
     "redux-thunk": "2.3.0",
     "resolve-url-loader": "2.3.0",
